+++ DDS-correctorParciales/src/correctorParciales/TablaConversion.java
@@ -1,9 +1,6 @@
 package correctorParciales;
 
-import Examen;
-
 import java.util.Collection;
-import java.util.List;
 import java.util.stream.Stream;
 
 public class TablaConversion implements MetodoCorreccion {
@@ -18,13 +15,13 @@ public class TablaConversion implements MetodoCorreccion {
 		return this.getParesConversion().stream();
 	}
 	
-	public void setNuevoParConversion(float unPuntaje, float unaNotaCorrespondiente) {
+	public void setNuevoParConversion(double unPuntaje, double unaNotaCorrespondiente) {
 		ParDeConversion unPar = new ParDeConversion(unPuntaje, unaNotaCorrespondiente);
 		this.getParesConversion().add(unPar);
 				
 	}
 
-	public float notaParaExamen(Examen unExamen){
+	public double notaParaExamen(Examen unExamen){
 		return this.getStreamDePares().filter(par-> par.coincideCon(unExamen.puntajeTotal())).findAny().get().getNotaCorrespondiente();
 		}
